requirements.txt
pandas
psycopg2-binary
sqlalchemy
azure-identity
azure-storage-file-datalake
azure-core
python-dotenv
